Extract HealthChecker
